+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - sitemap 加上 CDATA 防護 @esbb48
 
 ## 1.0.18 - 2017-08-07
+### Changed
+- 將圖片調整長 GoogleCDN @esbb48
+
 ### Fixed
 - 減少 firebase 不需要的套件的載入 @esbb48
 - 在 ie 時，執行推播會 404 @esbb48